+++ docs/changelog.md
@@ -32,6 +32,7 @@ Ends October 14, 2020
 - update hand card styles during mousemove events.
 - add canDrawCard to handleKeyboardEvent useCallback's dependency array.
 - add profiler script to be used with React's Profiler API.
+- add handleRender profiler function.
 
 ## Sprint 4
 Ends September 16, 2020